Trekk ut domene-typer fra reducern og put i shared-types mappe
