+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/familieforhold/personer/Barn.tsx
@@ -2,7 +2,7 @@ import { PersonInfoFnr } from './personinfo/PersonInfoFnr'
 import { PersonBorder, PersonHeader, PersonInfoWrapper } from '../styled'
 import { ChildIcon } from '~shared/icons/childIcon'
 import { PersonDetailWrapper, TypeStatusWrap } from '../../styled'
-import { IPdlPerson } from '~store/reducers/BehandlingReducer'
+import { IPdlPerson } from '~shared/types/Person'
 import { PersonInfoAdresse } from './personinfo/PersonInfoAdresse'
 import differenceInYears from 'date-fns/differenceInYears'
 import { hentAdresserEtterDoedsdato } from '~components/behandling/felles/utils'
