+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/soeknadsoversikt/soeknadoversikt/kommerBarnetTilgode/Adresse.tsx
@@ -1,4 +1,4 @@
-import { VurderingsResultat } from '~store/reducers/BehandlingReducer'
+import { VurderingsResultat } from '~shared/types/VurderingsResultat'
 import { OversiktElement } from '../OversiktElement'
 
 export const Adresse = ({ resultat, tekst }: { resultat: VurderingsResultat | undefined; tekst: string }) => {
