+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/saksliste.tsx
@@ -2,7 +2,7 @@ import { Table } from '@navikt/ds-react'
 import styled from 'styled-components'
 import { AarsaksTyper, IBehandlingsammendrag } from './typer'
 import { formaterStringDato, formaterEnumTilLesbarString } from '~utils/formattering'
-import { IBehandlingStatus } from '~store/reducers/BehandlingReducer'
+import { IBehandlingStatus } from '~shared/types/IDetaljertBehandling'
 
 const colonner = ['Opprettet', 'Type', 'Årsak', 'Status', 'Vedtaksdato', 'Resultat']
