+++ apps/etterlatte-saksbehandling-ui/client/src/shared/api/beregning.ts
@@ -1,4 +1,5 @@
 import { apiClient, ApiResponse } from '~shared/api/apiClient'
+import { Beregning } from '~shared/types/Beregning'
 
 export const hentBeregning = async (behandlingId: string): Promise<ApiResponse<Beregning>> => {
   return apiClient.get(`/beregning/${behandlingId}`)
@@ -6,28 +7,4 @@ export const hentBeregning = async (behandlingId: string): Promise<ApiResponse<B
 
 export const opprettEllerEndreBeregning = async (behandlingId: string): Promise<ApiResponse<Beregning>> => {
   return apiClient.post(`/beregning/${behandlingId}`, {})
-}
-
-export interface Beregning {
-  beregningId: string
-  behandlingId: string
-  beregningsperioder: Beregningsperiode[]
-  beregnetDato: string
-  grunnlagMetadata: GrunnlagMetadata
-}
-
-export interface GrunnlagMetadata {
-  sakId: string
-  versjon: number
-}
-
-export interface Beregningsperiode {
-  delytelsesId: string
-  type: string
-  datoFOM: string
-  datoTOM: string
-  utbetaltBeloep: number
-  soeskenFlokk: any[]
-  grunnbelopMnd: number
-  grunnbelop: number
 }
