+++ apps/etterlatte-saksbehandling-ui/client/src/shared/icons/gyldigIcon.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { VurderingsResultat } from '~store/reducers/BehandlingReducer'
+import { VurderingsResultat } from '~shared/types/VurderingsResultat'
 
 export const GyldighetIcon = (props: { status: VurderingsResultat; large?: boolean }) => {
   return (
