+++ apps/etterlatte-saksbehandling-ui/client/src/shared/icons/statusIcon.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { VurderingsResultat } from '~store/reducers/BehandlingReducer'
+import { VurderingsResultat } from '~shared/types/VurderingsResultat'
 
 export const StatusIcon = (props: { status: VurderingsResultat; large?: boolean; noLeftPadding?: boolean }) => {
   const symbol = hentSymbol()
