+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/Grunnlagsopplysning.ts
@@ -1,0 +1,6 @@
+export interface Grunnlagsopplysning<T> {
+  id: string
+  kilde: string
+  opplysningsType: string
+  opplysning: T
+}
