+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/IDetaljertBehandling.ts
@@ -1,0 +1,97 @@
+import { IVilkaarsvurdering } from '~shared/api/vilkaarsvurdering'
+import { Beregning } from '~shared/types/Beregning'
+import { IHendelse } from '~shared/types/IHendelse'
+import { JaNei } from '~shared/types/ISvar'
+import { KildeSaksbehandler } from '~shared/types/kilde'
+import { IKriterie } from '~shared/types/Kriterie'
+import { IFamilieforhold, IPdlPerson } from '~shared/types/Person'
+import { Utfall } from '~shared/types/Utfall'
+import { VurderingsResultat } from '~shared/types/VurderingsResultat'
+
+export interface IDetaljertBehandling {
+  id: string
+  sak: number
+  gyldighetsprøving?: IGyldighetResultat
+  kommerBarnetTilgode: IKommerBarnetTilgode | null
+  vilkårsprøving?: IVilkaarsvurdering
+  beregning?: Beregning
+  saksbehandlerId?: string
+  fastsatt: boolean
+  datoFattet?: string //kommer som Instant fra backend
+  datoAttestert?: string //kommer som Instant fra backend
+  attestant?: string
+  soeknadMottattDato: string
+  virkningstidspunkt: Virkningstidspunkt | null
+  status: IBehandlingStatus
+  hendelser: IHendelse[]
+  familieforhold?: IFamilieforhold
+  behandlingType: IBehandlingsType
+  søker?: IPdlPerson
+}
+
+export enum IBehandlingsType {
+  FØRSTEGANGSBEHANDLING = 'FØRSTEGANGSBEHANDLING',
+  REVURDERING = 'REVURDERING',
+  MANUELT_OPPHOER = 'MANUELT_OPPHOER',
+}
+
+export enum IBehandlingStatus {
+  OPPRETTET = 'OPPRETTET',
+  VILKAARSVURDERT = 'VILKAARSVURDERT',
+  BEREGNET = 'BEREGNET',
+  FATTET_VEDTAK = 'FATTET_VEDTAK',
+  ATTESTERT = 'ATTESTERT',
+  RETURNERT = 'RETURNERT',
+  IVERKSATT = 'IVERKSATT',
+  AVBRUTT = 'AVBRUTT',
+}
+
+export interface IGyldighetResultat {
+  resultat: VurderingsResultat | undefined
+  vurderinger: IGyldighetproving[]
+  vurdertDato: string
+}
+
+export enum GyldigFramsattType {
+  INNSENDER_ER_FORELDER = 'INNSENDER_ER_FORELDER',
+  HAR_FORELDREANSVAR_FOR_BARNET = 'HAR_FORELDREANSVAR_FOR_BARNET',
+  INGEN_ANNEN_VERGE_ENN_FORELDER = 'INGEN_ANNEN_VERGE_ENN_FORELDER',
+}
+
+export interface IGyldighetproving {
+  navn: GyldigFramsattType
+  resultat: VurderingsResultat
+  basertPaaOpplysninger: any
+}
+
+export interface IKommerBarnetTilgode {
+  svar: JaNei
+  begrunnelse: string
+  kilde: KildeSaksbehandler
+}
+
+export interface IVilkaarsproving {
+  navn: VilkaarsType
+  resultat: VurderingsResultat
+  utfall: Utfall | undefined
+  kriterier: IKriterie[]
+  vurdertDato: string
+}
+
+export interface Virkningstidspunkt {
+  dato: string
+  kilde: KildeSaksbehandler
+}
+
+export enum VilkaarsType {
+  SOEKER_ER_UNDER_20 = 'SOEKER_ER_UNDER_20',
+  DOEDSFALL_ER_REGISTRERT = 'DOEDSFALL_ER_REGISTRERT',
+  AVDOEDES_FORUTGAAENDE_MEDLEMSKAP = 'AVDOEDES_FORUTGAAENDE_MEDLEMSKAP',
+  BARNETS_MEDLEMSKAP = 'BARNETS_MEDLEMSKAP',
+  SAMME_ADRESSE = 'GJENLEVENDE_OG_BARN_SAMME_BOSTEDADRESSE',
+  BARN_BOR_PAA_AVDOEDES_ADRESSE = 'BARN_BOR_PAA_AVDOEDES_ADRESSE',
+  BARN_INGEN_OPPGITT_UTLANDSADRESSE = 'BARN_INGEN_OPPGITT_UTLANDSADRESSE',
+  SAKSBEHANDLER_RESULTAT = 'SAKSBEHANDLER_RESULTAT',
+  FORMAAL_FOR_YTELSEN = 'FORMAAL_FOR_YTELSEN',
+  SAKEN_KAN_BEHANDLES_I_SYSTEMET = 'SAKEN_KAN_BEHANDLES_I_SYSTEMET',
+}
