+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/Utfall.ts
@@ -1,0 +1,5 @@
+export enum Utfall {
+  OPPFYLT = 'OPPFYLT',
+  BEHANDLE_I_PSYS = 'BEHANDLE_I_PSYS',
+  TRENGER_AVKLARING = 'TRENGER_AVKLARING',
+}
