+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/VurderingsResultat.ts
@@ -1,0 +1,5 @@
+export enum VurderingsResultat {
+  OPPFYLT = 'OPPFYLT',
+  IKKE_OPPFYLT = 'IKKE_OPPFYLT',
+  KAN_IKKE_VURDERE_PGA_MANGLENDE_OPPLYSNING = 'KAN_IKKE_VURDERE_PGA_MANGLENDE_OPPLYSNING',
+}
